Update: + thêm chức năng ném bomb # chỉnh sửa lại đạn,

+++ Classes/GameObject.cpp
@@ -27,7 +27,7 @@ bool GameObject::init()
 
 
 
-int GameObject::getDamage()
+float GameObject::getDamage()
 {
 	return _Dmg;
 }
